spyfu/requirements.txt
requests==2.31.0
google-cloud-bigquery==3.11.4
pandas==2.1.4
pandas-gbq==0.19.2
numpy==1.26.4
functions-framework==3.4.0
pyyaml==6.0.1
